wired up option  for user to set their own defaults
